+++ cube/Entity/water.cpp
@@ -1,5 +1,6 @@
 #include "water.h"
 #include "material/materialpool.h"
+#include "texture/texturepool.h"
 Water::Water(float width, float height, float waterLevel, float scale)
 {
     initWaves();
@@ -45,9 +46,22 @@ Water::Water(float width, float height, float waterLevel, float scale)
     }
 
     m_mesh->setMaterial (MaterialPool::getInstance ()->createOrGetMaterial ("default"));
+
     m_mesh->finishWithoutNormal ();
+
     addMesh (m_mesh);
+
+    this->setShaderProgram (ShaderPool::getInstance()->get ("deferred_water"));
+
+    m_mesh->getMaterial ()->setNormalMap (TexturePool::getInstance ()->createOrGetTexture ("./res/texture/water/water_normal.jpg"));
+
+    m_mirrorRenderTarget = NULL;
+    m_mirrorCamera = new Camera();
+    m_mirrorCamera->setOrtho (-width*1.0/2,width*1.0/2,-height*1.0/2,height*1.0/2,0.01,100);
+    m_mirrorCamera->setRotation (QVector3D(90,0,0));
+    m_mirrorCamera->setName ("heihei");
 }
+
 float Water::width() const
 {
     return m_width;
@@ -57,6 +71,7 @@ void Water::setWidth(float width)
 {
     m_width = width;
 }
+
 float Water::height() const
 {
     return m_height;
@@ -101,7 +116,8 @@ float Water::evaluateWave(Wave w, QVector2D vPos, float fTime)
 
 void Water::adjustVertices()
 {
-    m_time +=0.1;
+    m_time +=0.05;
+    getShaderProgram ()->setUniformFloat ("g_time",m_time);
     for(int i =0;i<m_width;i++)
     {
         for(int j=0;j<m_height;j++)
@@ -115,15 +131,48 @@ void Water::adjustVertices()
             m_mesh->at (index)->position.setY (y);
         }
     }
+    if(m_mirrorRenderTarget)
+    {
+        getShaderProgram ()->setUniformInteger ("g_has_mirror",1);
+        getShaderProgram ()->setUniformInteger ("g_mirror_map",4);
+        glActiveTexture (GL_TEXTURE4);
+        glBindTexture (GL_TEXTURE_2D,m_mirrorRenderTarget->resultBuffer ()->texture ());
+    }
+    else
+    {
+        getShaderProgram ()->setUniformInteger ("g_has_mirror",0);
+    }
     m_mesh->finishWithoutNormal ();
 }
 
 void Water::initWaves()
 {
-    m_waves[0] = { 1.0f, 0.30f, 0.50f, QVector2D( -1.0f, 0.0f ) };
-   m_waves[1] =  { 2.0f, 0.25f, 1.30f, QVector2D( -0.7f, 0.7f ) };
-   m_waves[2] =  { .50f, .25f, 0.250f, QVector2D( 0.2f, 0.1f ) };
+    m_waves[0] = { 0.5f, 0.2f, 0.50f, QVector2D( -1.0f, 0.0f ) };
+   m_waves[1] =  { 1.0f, 0.17f, 1.30f, QVector2D( -0.7f, 0.7f ) };
+   m_waves[2] =  { .25f, 0.4f, 0.250f, QVector2D( 0.2f, 0.1f ) };
 }
+Camera *Water::mirrorCamera() const
+{
+    return m_mirrorCamera;
+}
+
+void Water::setMirrorCamera(Camera *mirrorCamera)
+{
+    m_mirrorCamera = mirrorCamera;
+}
+
+
+RenderTarget *Water::mirrorRenderTarget() const
+{
+    return m_mirrorRenderTarget;
+}
+
+void Water::setMirrorRenderTarget(RenderTarget *mirrorRenderTarget)
+{
+    m_mirrorRenderTarget = mirrorRenderTarget;
+    m_mirrorRenderTarget->setIgnoreEntity (this);
+}
+