+++ cube/Entity/water.h
@@ -5,6 +5,7 @@
 #include <QOpenGLFunctions>
 #include "geometry/mesh.h"
 #include <QVector2D>
+#include "base/RenderTarget.h"
 
 struct Wave{
     float    fFreq;    // 频率 (2PI / 波长)
@@ -17,6 +18,7 @@ class Water : public Entity,protected QOpenGLFunctions
 {
 public:
     Water(float width, float height, float waterLevel = 10 ,float waterScale = 1);
+
     float width() const;
     void setWidth(float width);
 
@@ -33,10 +35,19 @@ public:
     void setWaterLevel(float waterLevel);
 
     float evaluateWave(Wave w, QVector2D vPos, float fTime);
+
     virtual void adjustVertices();
+
+    RenderTarget *mirrorRenderTarget() const;
+    void setMirrorRenderTarget(RenderTarget *mirrorRenderTarget);
+
+    Camera *mirrorCamera() const;
+    void setMirrorCamera(Camera *mirrorCamera);
+
 private:
     void initWaves();
 private:
+    Camera * m_mirrorCamera;
     float m_time;
     Wave m_waves[3];
     TMesh * m_mesh;
@@ -44,6 +55,7 @@ private:
     float m_height;
     float m_waterScale;
     float m_waterLevel;
+    RenderTarget * m_mirrorRenderTarget;
 };
 
 #endif // WATER_H